расширение репозитория данных о пациентах
Команды для получения и добавления пациентов через mediator

Receiver Для получения данных о пациентах

+++ PatientsResolver.API.Entities/Patient.cs
@@ -17,7 +17,7 @@ namespace PatientsResolver.API.Entities
         [NotNull]
         [Key]
         [Column("Id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [NotNull]
         [Column("Name")]
@@ -29,6 +29,6 @@ namespace PatientsResolver.API.Entities
 
         [NotNull]
         [Column("MedicalHistoryNumber")]
-        public long MedicalHistoryNumber { get; set; }
+        public int MedicalHistoryNumber { get; set; }
     }
 }
